Fixed the closest side code and all the other crap related to calculating that.
