+++ tests/fluigi/pnr/sfc/test_compositecell.py
@@ -1,0 +1,21 @@
+def test_horizontal_spacer_column():
+    raise NotImplementedError()
+
+def test_vertical_spacer_column():
+    raise NotImplementedError()
+
+def test_get_spacer_size():
+    raise NotImplementedError()
+
+def test_generate_spacers():
+    raise NotImplementedError()
+
+def test_spread_ports():
+    raise NotImplementedError()
+
+def test_activate_ports():
+    raise NotImplementedError()
+
+def test_from_parchmint_component():
+    raise NotImplementedError()
+
