+++ tests/fluigi/pnr/sfc/test_spacer_insert.py
@@ -1,0 +1,12 @@
+def test_try_shift_left():
+    raise NotImplementedError()
+
+def test_try_shift_right():
+    raise NotImplementedError()
+
+def test_generate_bin_map():
+    raise NotImplementedError()
+
+def test_shift_furthest_feasible_point():
+    raise NotImplementedError()
+
